Refactor: code통합 및 swagger 통합
