+++ routes/apiDocs.js
@@ -2,7 +2,7 @@
 /**
  * @swagger
  * paths:
- *  /login:
+ *  /user/login:
  *    post:
  *      summary: 로그인
  *      tags:
@@ -16,12 +16,32 @@
  *          required: true
  *          type: string
  *          description : "email,password"
- *       
+ *
  *      responses:
  *       200:
  *        description: 로그인 성공
  *
-
  *
- *    
+ *
+ *  /user/login/a:
+ *    post:
+ *      summary: test
+ *      tags:
+ *      - test
+ *      description: test
+ *      produces:
+ *      - application/json
+ *      parameters:
+ *        - in: body
+ *          name: email
+ *          required: true
+ *          type: string
+ *          description : "email,password"
+ *
+ *      responses:
+ *       200:
+ *        description: 로그인 성공
+ *
+ *
+ *
  */
