+++ routes/index.js
@@ -1,23 +1,14 @@
 const express = require('express');
 const router = express.Router();
+const { specs, swaggerUi } = require('../modules/swagger');
 
-const loginRouter = require('./rogin');
+const usersRouter = require('./user');
 
-const { specs, swaggerUi } = require('../modules/swagger')
-const usersRouter = require('./users_signup')
-const excelDownLoadRoute = require("./excelDownLoad")
-const updatemgrRouter = require('./updatemgr')
+const excelDownLoadRoute = require('./excelDownLoad');
 
-router.use(loginRouter);
-router.use('/api-docs', swaggerUi.serve, swaggerUi.setup(specs));
 router.use('/user', usersRouter);
-router.use('/update', updatemgrRouter)
+router.use('/api-docs', swaggerUi.serve, swaggerUi.setup(specs));
+
 router.use('/excel', excelDownLoadRoute);
-
-
-
-
-
-
 
 module.exports = router;
