--- routes/rogin.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const usersController = require('../controllers/rogin');
-
-const router = express.Router();
-
-router.post('/login', usersController.login);
-
-module.exports = router;
